chore: re-organized file test by segmneted category
