--- test/infrastructure/provider/database_helper_test.mocks.dart
@@ -1,87 +1,0 @@
-// Mocks generated by Mockito 5.4.3 from annotations
-// in t_ddd_flutter/test/infrastructure/provider/database_helper_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i4;
-
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:sqflite_common_ffi/sqflite_ffi.dart' as _i2;
-import 'package:t_ddd_flutter/domain/entities/user.dart' as _i5;
-import 'package:t_ddd_flutter/infrastructure/provider/database_helper.dart'
-    as _i3;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: deprecated_member_use
-// ignore_for_file: deprecated_member_use_from_same_package
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-class _FakeDatabase_0 extends _i1.SmartFake implements _i2.Database {
-  _FakeDatabase_0(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
-/// A class which mocks [DatabaseHelper].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockDatabaseHelper extends _i1.Mock implements _i3.DatabaseHelper {
-  MockDatabaseHelper() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i4.Future<_i2.Database> get database => (super.noSuchMethod(
-        Invocation.getter(#database),
-        returnValue: _i4.Future<_i2.Database>.value(_FakeDatabase_0(
-          this,
-          Invocation.getter(#database),
-        )),
-      ) as _i4.Future<_i2.Database>);
-
-  @override
-  _i4.Future<dynamic> insert({
-    required String? table,
-    required Map<String, Object?>? data,
-  }) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #insert,
-          [],
-          {
-            #table: table,
-            #data: data,
-          },
-        ),
-        returnValue: _i4.Future<dynamic>.value(),
-      ) as _i4.Future<dynamic>);
-
-  @override
-  _i4.Future<_i5.User?> get({
-    required String? table,
-    required Object? id,
-  }) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #get,
-          [],
-          {
-            #table: table,
-            #id: id,
-          },
-        ),
-        returnValue: _i4.Future<_i5.User?>.value(),
-      ) as _i4.Future<_i5.User?>);
-}
