--- test/infrastructure/repository/user_repository_impl_test.mocks.dart
@@ -1,68 +1,0 @@
-// Mocks generated by Mockito 5.4.3 from annotations
-// in t_ddd_flutter/test/infrastructure/repository/user_repository_impl_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i3;
-
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:t_ddd_flutter/domain/entities/user.dart' as _i4;
-import 'package:t_ddd_flutter/infrastructure/repository/user_repository_impl.dart'
-    as _i2;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: deprecated_member_use
-// ignore_for_file: deprecated_member_use_from_same_package
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-/// A class which mocks [UserRepositoryImpl].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockUserRepositoryImpl extends _i1.Mock
-    implements _i2.UserRepositoryImpl {
-  MockUserRepositoryImpl() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i3.Future<_i4.User?> login(
-    String? username,
-    String? password,
-  ) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #login,
-          [
-            username,
-            password,
-          ],
-        ),
-        returnValue: _i3.Future<_i4.User?>.value(),
-      ) as _i3.Future<_i4.User?>);
-
-  @override
-  _i3.Future<_i4.User?> getUserById(int? id) => (super.noSuchMethod(
-        Invocation.method(
-          #getUserById,
-          [id],
-        ),
-        returnValue: _i3.Future<_i4.User?>.value(),
-      ) as _i3.Future<_i4.User?>);
-
-  @override
-  _i3.Future<bool> register(_i4.User? data) => (super.noSuchMethod(
-        Invocation.method(
-          #register,
-          [data],
-        ),
-        returnValue: _i3.Future<bool>.value(false),
-      ) as _i3.Future<bool>);
-}
